+++ src/shared.rs
@@ -13,7 +13,7 @@ use backend::connected_client::NetworkIdMap;
 use message::registry::RemoteMessageRegistry;
 use replication::{
     receive_markers::ReceiveMarkers, registry::ReplicationRegistry, rules::ReplicationRules,
-    signature::SignatureMap, track_mutate_messages::TrackMutateMessages,
+    signature::SignatureMap,
 };
 
 /// Initializes types, resources and events needed for both client and server.
@@ -103,7 +103,6 @@ impl Plugin for RepliconSharedPlugin {
             .init_state::<ServerState>()
             .init_resource::<ProtocolHasher>()
             .init_resource::<NetworkIdMap>()
-            .init_resource::<TrackMutateMessages>()
             .init_resource::<RepliconChannels>()
             .init_resource::<ReplicationRegistry>()
             .init_resource::<ReplicationRules>()