+++ src/shared/protocol.rs
@@ -108,11 +108,6 @@ impl ProtocolHasher {
         self.hash::<E>(ProtocolPart::IndependentEvent);
     }
 
-    pub(crate) fn track_mutate_messages(&mut self) {
-        debug!("enabling mutate message tracking");
-        ProtocolPart::TrackMutateMessages.hash(&mut self.0);
-    }
-
     fn hash<T>(&mut self, part: ProtocolPart) {
         part.hash(&mut self.0);
         any::type_name::<T>().hash(&mut self.0);
@@ -142,7 +137,6 @@ enum ProtocolPart {
     ServerEvent,
     IndependentMessage,
     IndependentEvent,
-    TrackMutateMessages,
     SharedMessage,
     SharedEvent,
 }
@@ -243,11 +237,10 @@ mod tests {
             hasher.add_server_event::<StructC>();
             hasher.add_client_message::<StructB>();
             hasher.add_client_event::<StructC>();
-            hasher.track_mutate_messages();
             hasher.add_custom(0usize);
         }
 
-        const EXPECTED: ProtocolHash = ProtocolHash(7833509759262497991);
+        const EXPECTED: ProtocolHash = ProtocolHash(10617357519866006183);
         assert_eq!(hasher1.finish(), EXPECTED);
         assert_eq!(hasher2.finish(), EXPECTED);
     }