+++ src/shared/replication.rs
@@ -8,7 +8,6 @@ pub mod registry;
 pub mod rules;
 pub mod signature;
 pub mod storage;
-pub mod track_mutate_messages;
 pub mod visibility;
 
 use bevy::prelude::*;