+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Removing `Replicated` from an entity now stops replication without despawning the entity on clients. Client-side despawns of `Remote` entities clean up their `ServerEntityMap` mappings. Despawning an entity still replicates as despawn.
+- Replace `TrackAppExt::track_mutate_messages` function with `ServerPlugin::track_mutate_messages` field. This setting no longer affects the protocol hash and can be set only on the server.
 
 ## [0.41.1] - 2026-06-24